Feat: Moves classes to individual files
