+++ tbtaf/rest/JobResultResponse.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import Dict, Optional, List
+
+class JobResultResponse(BaseModel):
+    job_id: str
+    results: Optional[List[Dict]] = None
+    summary: str
