+++ tbtaf/rest/SuitesResponse.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import Dict, Optional
+
+
+class SuitesResponse(BaseModel):
+    project_name: str
+    tags: Optional[Dict] = None
